+++ package.json
@@ -133,6 +133,7 @@
     "react-loadable": "^5.5.0",
     "react-number-format": "^4.9.3",
     "react-router-dom": "^6.3.0",
+    "react-spring-3d-carousel": "^1.3.2",
     "react-to-print": "^2.14.0",
     "react-transition-group": "^4.4.2",
     "shortid": "^2.2.16",
